refactor(api): claude_sessions 정적 라우트를 sources·discovery 모듈로 추출
external-paths 3개 + source-users 를 sources.py로, projects·empty/list·
ghost/list·DELETE /ghost 4개를 discovery.py로 이동. 핸들러 본문·경로·함수명
무변경 (원문 구간 diff 로 바이트 동일 확인).

__init__.py 의 include 순서에 두 모듈을 sessions·activity **앞에** 넣었다.
이 5개 라우트(external-paths·source-users·projects·processes·ghost)가
GET/DELETE /{session_id} 에 가려지는 대상이다.

라우트 테이블 characterization 테스트 통과 — HTTP 표면 25개 동일.
