+++ src/backend/api/claude_sessions/_legacy.py
@@ -4,9 +4,8 @@ import asyncio
 import logging
 from datetime import UTC, datetime
 
-from fastapi import APIRouter, Depends, HTTPException
-
-from api.deps import get_current_user_optional
+from fastapi import APIRouter
+from pydantic import BaseModel
 
 logger = logging.getLogger(__name__)
 
@@ -120,245 +119,6 @@ async def list_sessions(
         offset=offset,
         limit=limit,
     )
-
-
-# ========================================
-# Static routes (must come before /{session_id})
-# ========================================
-
-from pydantic import BaseModel
-
-
-class ExternalPathRequest(BaseModel):
-    """Request to add external path."""
-
-    path: str
-
-
-class ExternalPathResponse(BaseModel):
-    """Response for external path operations."""
-
-    success: bool
-    message: str
-    paths: list[str]
-
-
-@router.get("/external-paths", response_model=ExternalPathResponse)
-async def list_external_paths() -> ExternalPathResponse:
-    """List all external (non-default) projects paths.
-
-    Returns:
-        List of external paths currently configured
-    """
-    monitor = get_monitor()
-    paths = monitor.get_external_paths()
-
-    return ExternalPathResponse(
-        success=True,
-        message=f"Found {len(paths)} external path(s)",
-        paths=paths,
-    )
-
-
-@router.post("/external-paths", response_model=ExternalPathResponse)
-async def add_external_path(request: ExternalPathRequest) -> ExternalPathResponse:
-    """Add an external projects path at runtime.
-
-    Args:
-        request: Path to add
-
-    Returns:
-        Updated list of external paths
-    """
-    monitor = get_monitor()
-
-    if monitor.add_external_path(request.path):
-        return ExternalPathResponse(
-            success=True,
-            message=f"Added external path: {request.path}",
-            paths=monitor.get_external_paths(),
-        )
-    else:
-        raise HTTPException(
-            status_code=400,
-            detail=f"Path does not exist or is already added: {request.path}",
-        )
-
-
-@router.delete("/external-paths/{path_encoded}")
-async def remove_external_path(path_encoded: str) -> ExternalPathResponse:
-    """Remove an external projects path.
-
-    Note: Path should be URL-encoded (/ -> %2F)
-
-    Args:
-        path_encoded: URL-encoded path to remove
-
-    Returns:
-        Updated list of external paths
-    """
-    import urllib.parse
-
-    monitor = get_monitor()
-    path = urllib.parse.unquote(path_encoded)
-
-    if monitor.remove_external_path(path):
-        return ExternalPathResponse(
-            success=True,
-            message=f"Removed external path: {path}",
-            paths=monitor.get_external_paths(),
-        )
-    else:
-        raise HTTPException(
-            status_code=400,
-            detail=f"Path not found or is the default path: {path}",
-        )
-
-
-@router.get("/source-users")
-async def list_source_users() -> dict:
-    """List all unique source users from discovered sessions.
-
-    Returns:
-        List of unique usernames and current user
-    """
-    monitor = get_monitor()
-    users = monitor.get_unique_source_users()
-    current_user = monitor._get_current_user()
-
-    return {
-        "users": users,
-        "current_user": current_user,
-    }
-
-
-@router.get("/projects")
-async def list_projects(
-    current_user=Depends(get_current_user_optional),
-) -> dict:
-    """List project names for session filtering (접근 제어 적용).
-
-    접근 규칙:
-        - 시스템 admin: 모든 활성 프로젝트
-        - 조직 admin/owner: 자신의 조직 프로젝트 + 명시적 ProjectAccess
-        - 일반 member: 명시적 ProjectAccess만
-    """
-    import os
-
-    use_database = os.getenv("USE_DATABASE", "false").lower() == "true"
-
-    if use_database and current_user:
-        try:
-            from sqlalchemy import or_, select
-
-            from db.database import async_session_factory
-            from db.models import ProjectAccessModel, ProjectModel
-
-            is_admin = current_user.role == "admin" or current_user.is_admin
-
-            async with async_session_factory() as session:
-                if is_admin:
-                    result = await session.execute(
-                        select(ProjectModel.name)
-                        .where(ProjectModel.is_active == True)  # noqa: E712
-                        .order_by(ProjectModel.name)
-                    )
-                else:
-                    from api.projects import _get_admin_org_ids
-
-                    admin_org_ids = await _get_admin_org_ids(current_user)
-
-                    member_subq = (
-                        select(ProjectAccessModel.project_id)
-                        .where(ProjectAccessModel.user_id == current_user.id)
-                        .scalar_subquery()
-                    )
-
-                    if admin_org_ids:
-                        result = await session.execute(
-                            select(ProjectModel.name)
-                            .where(
-                                ProjectModel.is_active == True,  # noqa: E712
-                                or_(
-                                    ProjectModel.organization_id.in_(admin_org_ids),
-                                    ProjectModel.id.in_(member_subq),
-                                ),
-                            )
-                            .order_by(ProjectModel.name)
-                        )
-                    else:
-                        result = await session.execute(
-                            select(ProjectModel.name)
-                            .where(
-                                ProjectModel.is_active == True,  # noqa: E712
-                                ProjectModel.id.in_(member_subq),
-                            )
-                            .order_by(ProjectModel.name)
-                        )
-
-                project_names = [row[0] for row in result.fetchall()]
-                return {"projects": project_names}
-        except Exception as e:
-            logger.warning(f"DB project lookup failed, falling back to filesystem: {e}")
-
-    # Fallback: filesystem-based discovery
-    monitor = get_monitor()
-    projects = monitor.get_unique_projects()
-
-    return {
-        "projects": projects,
-    }
-
-
-@router.get("/empty/list")
-async def list_empty_sessions() -> dict:
-    """List all sessions with 0 messages.
-
-    Returns:
-        List of empty sessions
-    """
-    monitor = get_monitor()
-    empty_sessions = monitor.get_empty_sessions()
-
-    return {
-        "empty_count": len(empty_sessions),
-        "sessions": [s.model_dump() for s in empty_sessions],
-    }
-
-
-@router.get("/ghost/list")
-async def list_ghost_sessions() -> dict:
-    """List all ghost sessions (message_count > 0 but no real user/assistant messages).
-
-    These sessions have metadata entries but no actual conversation.
-
-    Returns:
-        List of ghost sessions
-    """
-    monitor = get_monitor()
-    ghost_sessions = monitor.get_ghost_sessions()
-
-    return {
-        "ghost_count": len(ghost_sessions),
-        "sessions": [s.model_dump() for s in ghost_sessions],
-    }
-
-
-@router.delete("/ghost")
-async def delete_ghost_sessions() -> dict:
-    """Delete all ghost sessions.
-
-    Returns:
-        List of deleted session IDs and count
-    """
-    monitor = get_monitor()
-    deleted_ids = monitor.delete_ghost_sessions()
-
-    return {
-        "success": True,
-        "deleted_count": len(deleted_ids),
-        "deleted_ids": deleted_ids,
-    }
 
 
 # ========================================
